Move livesync files from icenium-cli to common/appbuilder
Move livesync files from icenium-cli to common/appbuilder in order to support Livesync for Proton.
