--- lib/services/path-filtering.ts
@@ -1,60 +1,0 @@
-///<reference path="../.d.ts"/>
-"use strict";
-import minimatch = require("minimatch");
-
-export class PathFilteringService implements IPathFilteringService {
-
-	constructor(private $fs: IFileSystem) {	}
-
-	public getRulesFromFile(fullFilePath: string) : string[] {
-		let COMMENT_START = '#';
-		let rules: string[] = [];
-
-		try {
-			let fileContent = this.$fs.readText(fullFilePath).wait();
-			rules = _.reject(fileContent.split(/[\n\r]/),
-				(line: string) => line.length === 0 || line[0] === COMMENT_START);
-
-		} catch(e) {
-			if (e.code !== "ENOENT") { // file not found
-				throw e;
-			}
-		}
-
-		return rules;
-	}
-
-	public filterIgnoredFiles(files: string[], rules: string[], rootDir: string): string[]{
-		return _.reject(files, file => this.isFileExcluded(file, rules, rootDir));
-	}
-
-	public isFileExcluded(file: string, rules: string[], rootDir: string): boolean {
-		file = file.replace(rootDir, "").replace(new RegExp("^[\\\\|/]*"), "");
-		let fileMatched = true;
-		_.each(rules, rule => {
-			// minimatch treats starting '!' as pattern negation
-			// but we want the pattern matched and then do something else with the file
-			// therefore, we manually handle leading ! and \! and hide them from minimatch
-			let shouldInclude = rule[0] === '!';
-			if (shouldInclude) {
-				rule = rule.substr(1);
-				let ruleMatched = minimatch(file, rule, {nocase: true, dot: true});
-				if (ruleMatched) {
-					fileMatched = true;
-				}
-			} else {
-				let options = {nocase: true, nonegate: false, dot: true};
-				if (rule[0] === '\\' && rule[1] === '!') {
-					rule = rule.substr(1);
-					options.nonegate = true;
-				}
-				let ruleMatched = minimatch(file, rule, options);
-				fileMatched = fileMatched && !ruleMatched;
-			}
-		});
-
-		return !fileMatched;
-	}
-}
-
-$injector.register("pathFilteringService", PathFilteringService);
